% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{checkIDcondition}
\alias{checkIDcondition}
\title{check ID conditions}
\usage{
checkIDcondition(rf)
}
\arguments{
\item{rf}{nanoRF object}
}
\description{
check ID conditions
}
\examples{
# checkIDcondition(rf)
}
