% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{loadIDfile}
\alias{loadIDfile}
\title{load ID file}
\usage{
loadIDfile(rf, file)
}
\arguments{
\item{rf}{nanoRF object}

\item{file}{path to the ID file}
}
\value{
nanoRF object
}
\description{
load ID file
}
\examples{
# rf <- loadTrainingSet(rf, file = "H:/work/Project/061_20250124_Ohta_program/out/2025-01-24_check_original/load_training_set_ids20250110_T.R")
}
