% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{loadProteomicsData}
\alias{loadProteomicsData}
\title{loading Proteomics data}
\usage{
loadProteomicsData(rf, file, column.ratio = NULL, column.target = NULL)
}
\arguments{
\item{rf}{nanoRF object}

\item{file}{proteomics file path}

\item{column.ratio}{index of fold-change in input file}

\item{column.target}{index of important column. at least two columns from these should have values}
}
\value{
nanoRF object
}
\description{
loading Proteomics data
}
\examples{
# rf <- loadProteomicsData(rf, file = FILE_Proteomics_data, column.ratio = c(8:12), column.target = c(8, 9, 11, 12))
}
