% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{outputCutOff}
\alias{outputCutOff}
\title{output cutoff}
\usage{
outputCutOff(rf, file)
}
\arguments{
\item{rf}{nanoRF object}

\item{file}{output file}
}
\description{
output cutoff
}
\examples{
# outputCutoff(rf, file = "RF_cut_off.txt")
}
