% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{outputTable}
\alias{outputTable}
\title{output table}
\usage{
outputTable(rf, file)
}
\arguments{
\item{rf}{nanoRF object}

\item{file}{output file}
}
\description{
output table
}
\examples{
# outputTable(rf, file = "RF_score_result.txt")
}
