% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{plotPrediction}
\alias{plotPrediction}
\title{plot prediction graphs}
\usage{
plotPrediction(rf, output_dir)
}
\arguments{
\item{rf}{nanoRF object}

\item{output_dir}{output directory. path should have}
}
\description{
plot prediction graphs
}
\examples{
# plotPrediction(rf, output_dir = "output_directory/")
}
