% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{setNegative_vars}
\alias{setNegative_vars}
\title{change negative ID list}
\usage{
setNegative_vars(rf, ids)
}
\arguments{
\item{rf}{nanoRF object}

\item{ids}{vector of negative ID list}
}
\description{
change negative ID list
}
\examples{
# rf <- setNegative_vars(rf, ids = c("cytoskelton.raw.ids", "mitochondria.raw.ids", "membrane.raw.ids"))
}
