% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_by_step_functions.R
\name{setTarget_vars}
\alias{setTarget_vars}
\title{change target ID list}
\usage{
setTarget_vars(rf, ids)
}
\arguments{
\item{rf}{nanoRF object}

\item{ids}{vector of target IDs}
}
\description{
change target ID list
}
\examples{
# rf <- setTarget_vars(rf, ids = c("HMGA.raw.ids", "macroH2A.raw.ids"))
}
